독서 세션 종료후 navigation 처리

+++ BookKeep/BookKeep/Models/RealmBook.swift
@@ -55,10 +55,10 @@ class ReadSession: Object{
     @Persisted(primaryKey: true) var _id: ObjectId
     @Persisted var startTime: Date = .now
     @Persisted var endPage: Int = 0
-    @Persisted var duration: Int = 0
+    @Persisted var duration: Double = 0
     @Persisted(originProperty: "readSessions") var ofBook: LinkingObjects<RealmBook>
     
-    convenience init(startTime: Date, endPage: Int, duration: Int) {
+    convenience init(startTime: Date, endPage: Int, duration: Double) {
         self.init()
         self.startTime = startTime
         self.endPage = endPage
